requirements.txt
basicsr==1.4.2
beautifulsoup4==4.13.4
dominate==2.9.1
kornia==0.6.9
lpips==0.1.4
matplotlib==3.6.2
ml_collections==0.1.1
numpy==1.25.0
opencv_python==4.6.0.66
pandas==2.3.0
Pillow==9.3.0
Pillow==11.3.0
PyWavelets==1.4.1
PyYAML==6.0.1
PyYAML==6.0.2
Requests==2.32.4
scikit_image==0.19.3
scipy==1.16.0
seaborn==0.13.2
sync_batchnorm==0.0.1
timm==1.0.16
torch==1.13.0+cu116
torch_dct==0.1.6
torch_utils==0.1.2
torchvision==0.14.0+cu116
tqdm==4.67.1
visdom==0.2.4
wandb==0.16.2
